+++ spec/carve_spec.rb
@@ -107,6 +107,22 @@ RSpec.describe Rouge::Lexers::Carve do
       expect(lex('{#id .cls key="v"}')).to eq [['Name.Attribute', '{#id .cls key="v"}']]
     end
 
+    it 'reads an include directive as one token, selector included' do
+      # The selector is the point: `#intro` is genuinely tag syntax, and was
+      # coloured as one inside a path before the directive had a rule of its
+      # own (PART 9 section 19).
+      expect(lex('{{ ch.crv #intro }}')).to eq [['Name.Decorator', '{{ ch.crv #intro }}']]
+      expect(lex('{{ "a b.crv" }}')).to eq [['Name.Decorator', '{{ "a b.crv" }}']]
+    end
+
+    it 'leaves an include directive in a code span verbatim' do
+      expect(lex('`{{ x.crv }}`')).to eq [
+        ['Punctuation', '`'],
+        ['Literal.String.Backtick', '{{ x.crv }}'],
+        ['Punctuation', '`']
+      ]
+    end
+
     it 'reads a mention and a tag as one token each, sigil included' do
       expect(token_for('hi @user', '@user')).to eq 'Name.Variable.Magic'
       expect(token_for('a #tag', '#tag')).to eq 'Name.Variable.Instance'